third_party/blink/web_tests/accessibility/input-type-text-value-change-event.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<input id="text" type="text" value="before">

<script>
async_test(function(t)
{
    var target = document.getElementById("text");
    target.addEventListener("change", function() {
        assert_equals(target.value, "after");
        target.style.display = "none";
        t.done();
    }, false);

    var axTarget = accessibilityController.accessibleElementById("text");
    axTarget.setValue("after");
}, "changing the value of a text control fires a change event on it.");
</script>
